Add ability to start the DSP from bridge
Bridge program calls fork, in the child process it starts the qemu
DSP and in the parent process it starts the fuzzer and then they
start communicating.

Important thing to note here is that parent waits for 5 seconds
after it calls fork() so that child (qemu DSP) gets enough time
to initialize itself. This should be discussed and changed in
future.

Signed-off-by: Mohana Datta Yelugoti <ymdatta.work@gmail.com>