sof-afl-fuzzer: raise SIGABRT when sending message to DSP fails
The program raises SIGABRT signal when message sent to the DSP fails.
We call it a fail, when we are not able to receive a reply message
from the DSP in a certain time, after which we assume that DSP has
crashed.

Help option has also been added to the binary to give information on
how to run the fuzzer.

Signed-off-by: Mohana Datta Yelugoti <ymdatta.work@gmail.com>